+++ dashboard/src/index.tsx
@@ -1,9 +1,14 @@
 import * as React from "react";
 import * as ReactDOM from "react-dom";
+import { createAxiosInterceptors } from "shared/AxiosInstance";
+import { axios } from "./shared/Auth";
 
 import Root from "./containers/Root";
 import "./index.css";
+import store from "./store";
 // import registerServiceWorker from "./registerServiceWorker";
+
+createAxiosInterceptors(axios, store);
 
 ReactDOM.render(<Root />, document.getElementById("root") as HTMLElement);
