+++ dashboard/src/store/index.ts
@@ -1,4 +1,4 @@
-import { History } from "history";
+import createHistory from "history/createBrowserHistory";
 import { routerMiddleware } from "react-router-redux";
 import { applyMiddleware, createStore } from "redux";
 import { composeWithDevTools } from "redux-devtools-extension";
@@ -7,10 +7,8 @@ import thunkMiddleware from "redux-thunk";
 import rootReducer from "../reducers";
 import { IStoreState } from "../shared/types";
 
-const configureStore = (history: History) =>
-  createStore<IStoreState>(
-    rootReducer,
-    composeWithDevTools(applyMiddleware(thunkMiddleware, routerMiddleware(history))),
-  );
-
-export default configureStore;
+export const history = createHistory();
+export default createStore<IStoreState>(
+  rootReducer,
+  composeWithDevTools(applyMiddleware(thunkMiddleware, routerMiddleware(history))),
+);
